requirements.txt
flask
requests
pytest
python-dotenv
flask-cors
